moves joi schemas to utils
