+++ controllers/auth/logout.js
@@ -10,11 +10,7 @@ const logout = async (req, res) => {
   }
 
   await User.findByIdAndUpdate(_id, { token: null });
-  res.status(204).json({
-    status: "success",
-    code: 204,
-    message: "No content",
-  });
+  res.status(204);
 };
 
 module.exports = logout;
