+++ models/contact.js
@@ -1,7 +1,5 @@
 const { Schema, model } = require("mongoose");
 const { handleValidationErrors } = require("../errorHandlers");
-const Joi = require("joi");
-const myCustomJoi = Joi.extend(require("joi-phone-number"));
 
 const contactSchema = Schema(
   {
@@ -35,36 +33,6 @@ const contactSchema = Schema(
 
 contactSchema.post("save", handleValidationErrors);
 
-const addSchema = Joi.object({
-  name: Joi.string()
-    .min(3)
-    .max(30)
-    .required()
-    .messages({ "any.required": "Missing required field: Name" }),
-  phone: myCustomJoi
-    .string()
-    .required()
-    .phoneNumber()
-    .messages({ "any.required": "Missing required field: Phone" }),
-  email: Joi.string()
-    .email({
-      minDomainSegments: 2,
-      tlds: { allow: ["com", "net"] },
-    })
-    .required()
-    .messages({ "any.required": "Missing required field: Email" }),
-  favorite: Joi.boolean(),
-});
-
-const toggleFavoriteSchema = Joi.object({
-  favorite: myCustomJoi.boolean(),
-});
-
-const schemas = {
-  addSchema,
-  toggleFavoriteSchema,
-};
-
 const Contact = model("contact", contactSchema);
 
-module.exports = { Contact, schemas };
+module.exports = { Contact };
